Change link filtering
Instead of capturing lines with `http` in them, I'm changing the link
filtering to match lines with a period surrounded by word characters.
There have to be two characters on following the period so we don't pick
up phrases like `e.g.`. This should be fine because there are no single
character TLDs.